Working changes

+++ example/simple/client.c
@@ -45,11 +45,11 @@ int main() {
 
         char buf[BUF_SIZE];
         fprintf(stdout, "Received message:\n\n");
-        int len;
+        size_t len;
         while (netReceive(sock, buf, BUF_SIZE, &len) == ESUCCESS && len >= BUF_SIZE) {
-            fprintf(stdout, "%.*s", len, buf);
+            fprintf(stdout, "%.*s", (int) len, buf);
         }
-        fprintf(stdout, "%.*s\n\n", len, buf);
+        fprintf(stdout, "%.*s\n\n", (int) len, buf);
 
         netClose(&sock);
     }
